Using hash to encode farcaster types
