+++ src/monero/mod.rs
@@ -48,9 +48,6 @@ pub struct Ed25519;
 
 impl Curve for Monero {
     type Curve = Ed25519;
-    fn curve(&self) -> Self::Curve {
-        todo!()
-    }
 }
 
 impl Accordant for Monero {}
@@ -69,4 +66,29 @@ impl PrivateViewKey for Monero {
 
 impl Commitment for Monero {
     type Commitment = PubkeyHash;
+}
+
+use strict_encoding::{StrictEncode, StrictDecode};
+use monero::cryptonote::hash::Hash;
+impl StrictEncode for Ed25519 {
+    fn strict_encode<E: std::io::Write>(&self, mut e: E) -> Result<usize, strict_encoding::Error> {
+        let res = Hash::hash(&"Farcaster Ed25519".as_bytes()).to_bytes();
+        e.write(&res)?;
+        Ok(res.len())
+    }
+}
+
+impl StrictDecode for Ed25519 {
+    fn strict_decode<D: std::io::Read>(mut d: D) -> Result<Self, strict_encoding::Error> {
+        let mut buf = [0u8; 32];
+        d.read_exact(&mut buf)?;
+        let expected = Hash::hash(&"Farcaster Ed25519".as_bytes()).to_bytes();
+        if expected == buf {
+            Ok(Self)
+        } else {
+            Err(strict_encoding::Error::DataIntegrityError(
+                "Not Ed25519 type".to_string(),
+            ))
+        }
+    }
 }
